% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccAnalysis-Package.R
\docType{package}
\name{bccAnalysis}
\alias{bccAnalysis-package}
\alias{bccAnalysis}
\title{Analysis Toolbox for GRanges Data}
\description{
bccAnalysis : Package for ChIP-seq Analysis and Filtering
This package provides tools to import, create and analyse GRanges Objects
There are 6 main functions within this package: filter_by_pvalue, proportionOverlapTibble, overlapGRangeList, custom_venn_diagram, import_samples and compareGRangeWidth
}
\author{
\strong{Maintainer}: Jacob Martin \email{Jacob@martin-email.com}

}
