% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGRangeWidth.R
\name{compareGRangeWidth}
\alias{compareGRangeWidth}
\title{Width Comparison of 2 GRanges Objects}
\usage{
compareGRangeWidth(Obj1, Obj2)
}
\arguments{
\item{Obj1}{GRanges Object}

\item{Obj2}{GRanges Object}
}
\value{
Returns a plot that shows the width of Object 1 and Object 2. Plot = (x = Width of Peak) (y = Density of a particular width (How many peaks have that width))
}
\description{
Width Comparison of 2 GRanges Objects
}
