% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_venn_diagram.R
\name{custom_venn_diagram}
\alias{custom_venn_diagram}
\title{Comparing Overlap of two GRange Objects}
\usage{
custom_venn_diagram(conditionA, conditionB, DataA, DataB)
}
\arguments{
\item{conditionA}{String with the name of the DataA}

\item{conditionB}{String with the name of DataB}

\item{DataA}{GRanges Object}

\item{DataB}{GRanges Object}
}
\value{
GGplot | Venn Diagram showing overlaps between two GRanges Objects.
}
\description{
Comparing Overlap of two GRange Objects
}
\author{
Jacob Martin
}
