% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_pvalue.R
\name{filter_by_pvalue}
\alias{filter_by_pvalue}
\title{Filtering Function that filter by a custom p value and chromosome list}
\usage{
filter_by_pvalue(
  GRangeList,
  pvalue,
  chromosomes = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
)
}
\arguments{
\item{pvalue}{float value p value}

\item{chromosomes}{list of characters e.g. c("chr1", "chr2")}

\item{list}{List of GRanges Objects}
}
\value{
Returns a list of GRange Objects with a new column for normal p values. The list is sorted and edited based on the p value and chromosome inputs.
}
\description{
Filtering Function that filter by a custom p value and chromosome list
}
\author{
Jacob Martin
}
