% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_samples.R
\name{import_samples}
\alias{import_samples}
\title{Importing narrowPeak files and converting into a GRanges Object}
\usage{
import_samples(named_file_list)
}
\arguments{
\item{named_file_list}{A named list with file paths for narrowPeak files. These files will be imported as GRanges Objects.}
}
\value{
Returns a list of GRange Objects with a new column for normal p values. The list is sorted and edited based on the p value and chromosome inputs.
}
\description{
Importing narrowPeak files and converting into a GRanges Object
}
