% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapGRangeList.R
\name{overlapGRangeList}
\alias{overlapGRangeList}
\title{Filtering GRange Objects by number of overlaps with other samples. This function is used to merge multiple GRange Objects into one.}
\usage{
overlapGRangeList(grange_list, num_of_overlaps_required = 1)
}
\arguments{
\item{grange_list}{List of GRanges Objects that are similar or part of the same condition}

\item{num_of_overlaps_required}{Integer value that determines how many overlaps a peak must have with the other GRanges objects in order to be outputed by the function}
}
\value{
Returns a single GRange Object that is merged and filtered by overlaps.
}
\description{
Filtering GRange Objects by number of overlaps with other samples. This function is used to merge multiple GRange Objects into one.
}
