% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionOverlapTibble.R
\name{proportionOverlapTibble}
\alias{proportionOverlapTibble}
\title{Filtering Function that filter by a custom p value and chromosome list}
\usage{
proportionOverlapTibble(
  object1,
  object2,
  overlapobject1,
  overlapobject2,
  object1_unique,
  object2_unique
)
}
\arguments{
\item{object1}{GRanges Object}

\item{object2}{GRanges Object}

\item{overlapobject1}{The overlap G ranges object of object1}

\item{overlapobject2}{The overlap G ranges object of object2}

\item{object1_unique}{The unique peaks as a G ranges object of object1}

\item{object2_unique}{The unique peaks as a G ranges object of object2}
}
\value{
Returns a tibble that shows: Proportion of Overlap, TotalPeak coverage, Total number of unique bases, number of overlap peaks and percentage overlap of peaks for object 1 and 2.
}
\description{
Filtering Function that filter by a custom p value and chromosome list
}
