% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{ShinyApp}
\alias{ShinyApp}
\title{Shiny App for Chip-Seq Analysis}
\usage{
ShinyApp(named_file_list, height = 850)
}
\arguments{
\item{named_file_list}{A named list with file paths for narrowPeak files. These files will be imported as GRanges Objects.}

\item{height}{Controls Browser window size on first launch}
}
\value{
Returns an Analysis App. Outputs: Overlaps, Unique peak annotations, width analysis and proportion overlap table
}
\description{
Shiny App for Chip-Seq Analysis
}
\author{
Jacob Martin
}
