% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccAnalysis-Package.R
\name{bccAnalysis}
\alias{bccAnalysis}
\title{bccAnalysis: ChIP-Seq Peak Analysis Tools}
\description{
Analysis tools for ChIP-Seq peak data including overlap analysis,
peak annotation, and visualization with Shiny applications.
}
\details{
The bccAnalysis package provides tools for processing GRanges objects from
narrowPeak files, performing overlap analysis, peak annotation using ChIPseeker,
and interactive Shiny applications for visualization.
}
\seealso{
\code{\link{ShinyApp}}
}
\author{
Jacob Martin
}
